+++ lib/mshoplib/src/MShop/Order/Item/Base/Product/Attribute/Iface.php
@@ -92,17 +92,17 @@ interface Iface
 	/**
 	 * Returns the quantity of the product attribute.
 	 *
-	 * @return int Quantity of the product attribute
+	 * @return float Quantity of the product attribute
 	 */
-	public function getQuantity() : int;
+	public function getQuantity() : float;
 
 	/**
 	 * Sets the quantity of the product attribute.
 	 *
-	 * @param int $value Quantity of the product attribute
+	 * @param float $value Quantity of the product attribute
 	 * @return \Aimeos\MShop\Order\Item\Base\Product\Attribute\Iface Order base product attribute item for chaining method calls
 	 */
-	public function setQuantity( int $value ) : \Aimeos\MShop\Order\Item\Base\Product\Attribute\Iface;
+	public function setQuantity( float $value ) : \Aimeos\MShop\Order\Item\Base\Product\Attribute\Iface;
 
 	/**
 	 * Copys all data from a given attribute item.
