+++ lib/mshoplib/src/MShop/Order/Item/Base/Service/Attribute/Standard.php
@@ -197,9 +197,9 @@ class Standard
 	/**
 	 * Returns the quantity of the service attribute.
 	 *
-	 * @return integer Quantity of the service attribute
+	 * @return float Quantity of the service attribute
 	 */
-	public function getQuantity() : int
+	public function getQuantity() : float
 	{
 		return $this->get( 'order.base.service.attribute.quantity', 1 );
 	}
@@ -208,10 +208,10 @@ class Standard
 	/**
 	 * Sets the quantity of the service attribute.
 	 *
-	 * @param int $value Quantity of the service attribute
+	 * @param float $value Quantity of the service attribute
 	 * @return \Aimeos\MShop\Order\Item\Base\Product\Attribute\Iface Order base service attribute item for chaining method calls
 	 */
-	public function setQuantity( int $value ) : \Aimeos\MShop\Order\Item\Base\Service\Attribute\Iface
+	public function setQuantity( float $value ) : \Aimeos\MShop\Order\Item\Base\Service\Attribute\Iface
 	{
 		return $this->set( 'order.base.service.attribute.quantity', $value );
 	}
