+++ lib/mshoplib/src/MShop/Price/Item/Standard.php
@@ -139,21 +139,21 @@ class Standard extends Base
 	/**
 	 * Returns the quantity the price is valid for.
 	 *
-	 * @return int Quantity
+	 * @return float Quantity
 	 */
-	public function getQuantity() : int
+	public function getQuantity() : float
 	{
-		return $this->get( 'price.quantity', 1 );
+		return (float) $this->get( 'price.quantity', 1 );
 	}
 
 
 	/**
 	 * Sets the quantity the price is valid for.
 	 *
-	 * @param int $quantity Quantity
+	 * @param float $quantity Quantity
 	 * @return \Aimeos\MShop\Price\Item\Iface Price item for chaining method calls
 	 */
-	public function setQuantity( int $quantity ) : \Aimeos\MShop\Price\Item\Iface
+	public function setQuantity( float $quantity ) : \Aimeos\MShop\Price\Item\Iface
 	{
 		return $this->set( 'price.quantity', $quantity );
 	}
@@ -396,10 +396,10 @@ class Standard extends Base
 	 * Add the given price to the current one.
 	 *
 	 * @param \Aimeos\MShop\Price\Item\Iface $item Price item which should be added
-	 * @param int $quantity Number of times the Price should be added
+	 * @param float $quantity Number of times the Price should be added
 	 * @return \Aimeos\MShop\Price\Item\Iface Price item for chaining method calls
 	 */
-	public function addItem( \Aimeos\MShop\Price\Item\Iface $item, int $quantity = 1 ) : \Aimeos\MShop\Price\Item\Iface
+	public function addItem( \Aimeos\MShop\Price\Item\Iface $item, float $quantity = 1 ) : \Aimeos\MShop\Price\Item\Iface
 	{
 		if( $item->getCurrencyId() != $this->getCurrencyId() )
 		{
